requirements.txt
pyautogui
numpy
pillow
keyboard
opencv-python
pyinstaller